N-01: rename uint to uint256

+++ contracts/bls/BLSAccountFactory.sol
@@ -26,7 +26,7 @@ contract BLSAccountFactory {
      * This method returns an existing account address so that entryPoint.getSenderAddress() would work even after account creation
      * Also note that out BLSSignatureAggregator requires that the public-key is the last parameter
      */
-    function createAccount(uint salt, uint256[4] memory aPublicKey) public returns (BLSAccount) {
+    function createAccount(uint256 salt, uint256[4] memory aPublicKey) public returns (BLSAccount) {
 
         address addr = getAddress(salt, aPublicKey);
         uint codeSize = addr.code.length;
@@ -42,7 +42,7 @@ contract BLSAccountFactory {
     /**
      * calculate the counterfactual address of this account as it would be returned by createAccount()
      */
-    function getAddress(uint salt, uint256[4] memory aPublicKey) public view returns (address) {
+    function getAddress(uint256 salt, uint256[4] memory aPublicKey) public view returns (address) {
         return Create2.computeAddress(bytes32(salt), keccak256(abi.encodePacked(
                 type(ERC1967Proxy).creationCode,
                 abi.encode(
